+++ test/unit/CMakeLists.txt
@@ -1,7 +1,5 @@
 add_scylla_test(bptree_compaction_test
   KIND UNIT)
-add_scylla_test(btree_compaction_test
-  KIND UNIT)
 add_scylla_test(cross_shard_barrier_test
   KIND UNIT)
 add_scylla_test(lsa_sync_eviction_test
